requirements.txt
alembic==1.13.2
amqp==5.2.0
aniso8601==9.0.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==24.2.0
blinker==1.8.2
cffi==1.17.1
click==8.1.7
Flask==3.0.3
flask-restx==1.3.0
greenlet==3.1.0
gunicorn==23.0.0
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.1
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.3
python-environ==0.4.54
pytz==2024.2
referencing==0.35.1
rpds-py==0.20.0
SQLAlchemy==2.0.34
typing_extensions==4.12.2
vine==5.1.0
Werkzeug==3.0.4
